Fix int-truncation in test_merkle and test_dict Debug builds
uc2_merkle_root() and uc2_dict_id() return uint64_t; the int _r
temporaries from 97e05ad's NDEBUG fix truncated the high 32 bits.
Under Release the assertion was stripped, hiding the bug; under
Debug the truncated _r never matched the second uint64_t call.

+++ tests/src/test_dict.c
@@ -27,7 +27,7 @@ static void test_id_deterministic(void)
 	struct uc2_dict d1, d2;
 	uc2_dict_create(&d1, data, sizeof data);
 	uc2_dict_create(&d2, data, sizeof data);
-	{ int _r = uc2_dict_id(&d1); (void)_r; assert(_r == uc2_dict_id(&d2)); }
+	{ uint64_t _r = uc2_dict_id(&d1); (void)_r; assert(_r == uc2_dict_id(&d2)); }
 	assert(uc2_dict_match(&d1, &d2));
 	uc2_dict_free(&d1);
 	uc2_dict_free(&d2);
